requirements.txt
flask==3.0.2
python-dotenv==1.0.1
supabase==2.3.5
openai==1.12.0
requests==2.31.0
gunicorn==21.2.0
flask-login==0.6.3
python-slugify==8.0.4
Pillow==10.2.0 